refactor(feature-analyser): Rename file to feature_analyser for consistency
Additionally, the folder structure for dataset and feature_analyser has been removed to prevent issues with import paths
